Add files via upload
And working on creating a 3d level design in unreal for a AAA title RPG game  While creating this level i have place enemy positions and cover box so that player use it for stealth. Red blocks represent enemies green one's are for cover blocks and Yellow cone is for the level objective.